backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai==1.3.0
python-dotenv==1.0.0
aiofiles==23.2.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pillow==10.1.0
ffmpeg-python==0.2.0
supabase==2.0.0
PyJWT==2.8.0
opencv-python==4.8.1.78
numpy==1.24.3

